+++ src/strigui/button.clj
@@ -2,13 +2,13 @@
   (:require [strigui.box :as b]
             [strigui.widget :as wdg]))
 
-(defrecord Button [name value args]
+(defrecord Button [name value props]
   wdg/Widget
-  (coord [this canvas] (apply b/box-coord [canvas (:value this) (:args this)]))
-  (defaults [this] (assoc-in this [:args :has-border?] true))
+  (coord [this canvas] (apply b/box-coord [canvas (:value this) (:props this)]))
+  (defaults [this] (assoc-in this [:props :has-border?] true))
   (before-drawing [this] this)
   (draw [this canvas]
-        (b/box-draw canvas (:value this) (:args this)))
+        (b/box-draw canvas (:value this) (:props this)))
   (after-drawing [this] this))
 
 (defmethod wdg/widget-event [strigui.button.Button :key-pressed]