+++ src/strigui/list.clj
@@ -13,24 +13,24 @@
 
 (defn get-index-at 
   [widget y]
-  (let [index (int (/ (- y (-> widget :args :y)) item-height))]
+  (let [index (int (/ (- y (-> widget :props :y)) item-height))]
     index))
 
 (defn draw-list
-  [canvas items args]
+  [canvas items props]
     (loop [items (filter :visible? items)
            index 0]
       (when (seq items)
-        (let [color (:color args)
+        (let [color (:color props)
               item (first items)
               color (if (or (:hovered? item) (:selected? item)) (update color :background make-darker-or-brighter) color)]
           (b/box-draw canvas (-> items first :value) (merge 
-                                                      {:y (+ (* index item-height) (:y args)) :max-width (- (:width args) item-width-right-margin) :color color}
-                                                      (select-keys args [:width :x])))
+                                                      {:y (+ (* index item-height) (:y props)) :max-width (- (:width props) item-width-right-margin) :color color}
+                                                      (select-keys props [:width :x])))
           (when (:selected? item)
             (c/draw-> canvas
-                      (c/line (inc (:x args)) (+ (* index item-height) (:y args)) 
-                              (inc (:x args)) (+ (* (inc index) item-height) (:y args)) 
+                      (c/line (inc (:x props)) (+ (* index item-height) (:y props)) 
+                              (inc (:x props)) (+ (* (inc index) item-height) (:y props)) 
                               (get color :text java.awt.Color/green) 3))))
         (recur (rest items) (inc index)))))
 
@@ -41,16 +41,16 @@
 
 (defn make-visible 
   ([widget]
-  (let [max-visible-items (Math/floor (/ (-> widget :args :height) item-height))
+  (let [max-visible-items (Math/floor (/ (-> widget :props :height) item-height))
         items (:items widget)
         widgets-visible (mapv #(merge % {:visible? true}) (take max-visible-items items))
         widgets-not-visible (take-last (- (count items) max-visible-items) items)]
     (assoc widget :items (apply conj widgets-visible widgets-not-visible))))
   ([widget y]
-   (let [max-visible-items (int (/ (-> widget :args :height) item-height))
+   (let [max-visible-items (int (/ (-> widget :props :height) item-height))
          items (:items widget)
          cnt-items (count items)
-         index (int (/ (int (- y (-> widget :args :y))) item-height))
+         index (int (/ (int (- y (-> widget :props :y))) item-height))
          until (+ index max-visible-items)
          diff (if (> until cnt-items) (- until cnt-items) 0)
          index (int (if (< (- index diff) 0) 0 (- index diff)))
@@ -64,17 +64,17 @@
      (assoc widget :items items))))
 
 
-(defrecord List [name items args]
+(defrecord List [name items props]
   wdg/Widget
-  (coord [this _] (let [{:keys [x y width height]} (:args this)]
+  (coord [this _] (let [{:keys [x y width height]} (:props this)]
                          [x y width height]))
   (defaults [this] (make-visible this))
   (before-drawing [this] (if (< (count (filter :visible? (:items this)))
-                                (Math/floor (/ (-> this :args :height) item-height)))
+                                (Math/floor (/ (-> this :props :height) item-height)))
                            (make-visible this)
                            this))
   (draw [this canvas]
-        (let [{:keys [x y width height color] :as args} (:args this)
+        (let [{:keys [x y width height color] :as props} (:props this)
               bar-x (+ x (- width item-width-right-margin))
               items (:items this)
               items-visible (filter :visible? items)
@@ -86,7 +86,7 @@
                     (c/rect x y width height (:background color) false)
                     (c/rect bar-x y item-width-right-margin height (make-darker-or-brighter (:background color)) true)
                     (c/rect bar-x bar-y item-width-right-margin bar-scroll-height (make-darker-or-brighter (:text color)) true))
-          (draw-list canvas items args)
+          (draw-list canvas items props)
           this))
   (after-drawing [this] this))
 
@@ -104,7 +104,7 @@
 
 (defmethod wdg/widget-event [strigui.list.List :mouse-clicked]
  [_ canvas widgets widget x y]
-  (let [item-border-x (+ (-> widget :args :x) (- (-> widget :args :width) item-width-right-margin))]
+  (let [item-border-x (+ (-> widget :props :x) (- (-> widget :props :width) item-width-right-margin))]
     (if (< item-border-x x)
       (update widgets (:name widget) make-visible y)
       (update widgets (:name widget) activate! y :selected?))))
@@ -115,7 +115,7 @@
 
 (defmethod wdg/widget-event [strigui.list.List :mouse-dragged]
   [_ canvas widgets widget x y _ _]
-  (let [item-border-x (+ (-> widget :args :x) (- (-> widget :args :width) item-width-right-margin))]
+  (let [item-border-x (+ (-> widget :props :x) (- (-> widget :props :width) item-width-right-margin))]
     (if (< item-border-x x)
       (update widgets (:name widget) make-visible y)
       widgets)))